+++ pkg/services/app/install.go
@@ -1,0 +1,9 @@
+package app
+
+import "fmt"
+
+type InstallParams struct{}
+
+func (a *App) Install(params InstallParams) error {
+	return fmt.Errorf("app.Install not implemented")
+}
